done the raycast part using our old repo as a reference , there is small issues on rendering

+++ src/raycast/calculate_wall_dist.c
@@ -6,7 +6,7 @@
 /*   By: jait-chd <jait-chd@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/08 16:26:16 by jait-chd          #+#    #+#             */
-/*   Updated: 2025/12/14 09:51:02 by jait-chd         ###   ########.fr       */
+/*   Updated: 2025/12/14 10:21:47 by jait-chd         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	calculate_wall_distance(t_data *data, t_ray *ray)
 {
 	if (ray->side == 0)
-		ray->perp_wall_dist = (ray->map_x - data->player.x + (1 - ray->step_x) / 2) / ray->dir_x;
+		ray->perp_wall_dist = ((double)ray->map_x - data->player.x + (1.0 - (double)ray->step_x) / 2.0) / ray->dir_x;
 	else
-		ray->perp_wall_dist = (ray->map_y - data->player.y + (1 - ray->step_y) / 2) / ray->dir_y;
+		ray->perp_wall_dist = ((double)ray->map_y - data->player.y + (1.0 - (double)ray->step_y) / 2.0) / ray->dir_y;
 }
